Main Menu
This is the main menu I worked on in order for data of the username and password to be saved to a text file so when the user logs in it will automatically be saved and the program will recognize whether the username and password match or not. I did not add the having users enter their classes because I figured you guys would be able to add that part to the code.